Add HTML to RSX converter
I got irritated converting HTML to RSX, so I made a little tool to do it for me. Might be a good fit for here?

I am unsure if this is of interest, it's also deployed on cargo-shuttle, which I haven't seen a guide for, so might be interesting of an example of that.